requirements.txt
django-sekizai
python-dateutil

# Additional requirements go here

